Move OpsDB property listing to athenz API (I know, I know ...)
