+++ controller-server/src/test/java/com/yahoo/vespa/hosted/controller/restapi/application/ApplicationApiTest.java
@@ -139,9 +139,6 @@ public class ApplicationApiTest extends ControllerContainerTest {
         // GET API root
         tester.assertResponse(request("/application/v4/", GET).userIdentity(USER_ID),
                               new File("root.json"));
-        // GET OpsDB properties
-        tester.assertResponse(request("/application/v4/property/", GET).userIdentity(USER_ID),
-                              new File("property-list.json"));
         // POST (add) a tenant without property ID
         tester.assertResponse(request("/application/v4/tenant/tenant1", POST)
                                       .userIdentity(USER_ID)
