+++ controller-server/src/test/java/com/yahoo/vespa/hosted/controller/restapi/application/responses/root.json
@@ -5,9 +5,6 @@
     },
     {
       "url":"http://localhost:8080/application/v4/tenant/"
-    },
-    {
-      "url":"http://localhost:8080/application/v4/property/"
     }
   ]
 }
